+++ tools/setup_sdcard.sh
@@ -1106,18 +1106,20 @@ populate_rootfs () {
 	fi
 
 	if [ -f ${TEMPDIR}/disk/etc/dnsmasq.conf ] ; then
-		wfile="/etc/dnsmasq.d/usb0-dhcp"
-		echo "#disable DNS by setting port to 0" > ${TEMPDIR}/disk${wfile}
-		echo "port=0" >> ${TEMPDIR}/disk${wfile}
-		echo "" >> ${TEMPDIR}/disk${wfile}
-		echo "interface=usb0" >> ${TEMPDIR}/disk${wfile}
-		echo "#one address range" >> ${TEMPDIR}/disk${wfile}
-		echo "dhcp-range=192.168.7.1,192.168.7.1" >> ${TEMPDIR}/disk${wfile}
-		echo "" >> ${TEMPDIR}/disk${wfile}
-		echo "dhcp-option=3" >> ${TEMPDIR}/disk${wfile}
-		echo "except-interface=lo" >> ${TEMPDIR}/disk${wfile}
-		echo "listen-address=192.168.7.2" >> ${TEMPDIR}/disk${wfile}
-		echo "bind-interfaces" >> ${TEMPDIR}/disk${wfile}
+		if [ ! -f ${TEMPDIR}/etc/dnsmasq.d/usb0-dhcp ] ; then
+			wfile="/etc/dnsmasq.d/usb0-dhcp"
+			echo "#disable DNS by setting port to 0" > ${TEMPDIR}/disk${wfile}
+			echo "port=0" >> ${TEMPDIR}/disk${wfile}
+			echo "" >> ${TEMPDIR}/disk${wfile}
+			echo "interface=usb0" >> ${TEMPDIR}/disk${wfile}
+			echo "#one address range" >> ${TEMPDIR}/disk${wfile}
+			echo "dhcp-range=192.168.7.1,192.168.7.1" >> ${TEMPDIR}/disk${wfile}
+			echo "" >> ${TEMPDIR}/disk${wfile}
+			echo "dhcp-option=3" >> ${TEMPDIR}/disk${wfile}
+			echo "except-interface=lo" >> ${TEMPDIR}/disk${wfile}
+			echo "listen-address=192.168.7.2" >> ${TEMPDIR}/disk${wfile}
+			echo "bind-interfaces" >> ${TEMPDIR}/disk${wfile}
+		fi
 	fi
 
 	if [ ! -f ${TEMPDIR}/disk/opt/scripts/boot/generic-startup.sh ] ; then
